+++ crates/block-producer/src/batch_builder/batch.rs
@@ -1,9 +1,11 @@
+use std::collections::BTreeSet;
+
 use miden_objects::{
     accounts::AccountId,
     batches::BatchNoteTree,
     block::BlockAccountUpdate,
     crypto::hash::blake::{Blake3Digest, Blake3_256},
-    notes::Nullifier,
+    notes::{NoteId, Nullifier},
     transaction::{OutputNote, TransactionId, TxAccountUpdate},
     Digest, MAX_NOTES_PER_BATCH,
 };
@@ -24,9 +26,10 @@ pub type BatchId = Blake3Digest<32>;
 pub struct TransactionBatch {
     id: BatchId,
     updated_accounts: Vec<(TransactionId, TxAccountUpdate)>,
+    unauthenticated_input_notes: BTreeSet<NoteId>,
     produced_nullifiers: Vec<Nullifier>,
-    created_notes_smt: BatchNoteTree,
-    created_notes: Vec<OutputNote>,
+    output_notes_smt: BatchNoteTree,
+    output_notes: Vec<OutputNote>,
 }
 
 impl TransactionBatch {
@@ -44,38 +47,61 @@ impl TransactionBatch {
     pub fn new(txs: Vec<ProvenTransaction>) -> Result<Self, BuildBatchError> {
         let id = Self::compute_id(&txs);
 
-        // TODO: we need to handle a possibility that a batch contains multiple transactions against
-        //       the same account (e.g., transaction `x` takes account from state `A` to `B` and
-        //       transaction `y` takes account from state `B` to `C`). These will need to be merged
-        //       into a single "update" `A` to `C`.
-        let updated_accounts =
-            txs.iter().map(|tx| (tx.id(), tx.account_update().clone())).collect();
+        let mut updated_accounts = vec![];
+        let mut produced_nullifiers = vec![];
+        let mut unauthenticated_input_notes = BTreeSet::new();
+        for tx in &txs {
+            // TODO: we need to handle a possibility that a batch contains multiple transactions against
+            //       the same account (e.g., transaction `x` takes account from state `A` to `B` and
+            //       transaction `y` takes account from state `B` to `C`). These will need to be merged
+            //       into a single "update" `A` to `C`.
+            updated_accounts.push((tx.id(), tx.account_update().clone()));
 
-        let produced_nullifiers = txs
+            for note in tx.input_notes() {
+                produced_nullifiers.push(note.nullifier());
+                if let Some(header) = note.header() {
+                    if !unauthenticated_input_notes.insert(header.id()) {
+                        return Err(BuildBatchError::DuplicateUnauthenticatedNote(
+                            header.id(),
+                            txs,
+                        ));
+                    }
+                }
+            }
+        }
+
+        // Populate batch output notes, filtering out unauthenticated notes consumed in the same batch.
+        // Consumed notes are also removed from the unauthenticated input notes set in order to avoid
+        // consumption of notes with the same ID by one single input.
+        //
+        // One thing to note:
+        // This still allows transaction `A` to consume an unauthenticated note `x` and output note `y`
+        // and for transaction `B` to consume an unauthenticated note `y` and output note `x`
+        // (i.e., have a circular dependency between transactions), but this is not a problem.
+        let output_notes: Vec<_> = txs
             .iter()
-            .flat_map(|tx| tx.input_notes().iter())
-            .map(|note| note.nullifier())
+            .flat_map(|tx| tx.output_notes().iter())
+            .filter(|&note| !unauthenticated_input_notes.remove(&note.id()))
+            .cloned()
             .collect();
 
-        let created_notes: Vec<_> =
-            txs.iter().flat_map(|tx| tx.output_notes().iter()).cloned().collect();
-
-        if created_notes.len() > MAX_NOTES_PER_BATCH {
-            return Err(BuildBatchError::TooManyNotesCreated(created_notes.len(), txs));
+        if output_notes.len() > MAX_NOTES_PER_BATCH {
+            return Err(BuildBatchError::TooManyNotesCreated(output_notes.len(), txs));
         }
 
         // TODO: document under what circumstances SMT creating can fail
-        let created_notes_smt = BatchNoteTree::with_contiguous_leaves(
-            created_notes.iter().map(|note| (note.id(), note.metadata())),
+        let output_notes_smt = BatchNoteTree::with_contiguous_leaves(
+            output_notes.iter().map(|note| (note.id(), note.metadata())),
         )
         .map_err(|e| BuildBatchError::NotesSmtError(e, txs))?;
 
         Ok(Self {
             id,
             updated_accounts,
+            unauthenticated_input_notes,
             produced_nullifiers,
-            created_notes_smt,
-            created_notes,
+            output_notes_smt,
+            output_notes,
         })
     }
 
@@ -108,19 +134,24 @@ impl TransactionBatch {
         })
     }
 
+    /// Returns unauthenticated input notes set consumed by the transactions in this batch.
+    pub fn unauthenticated_input_notes(&self) -> &BTreeSet<NoteId> {
+        &self.unauthenticated_input_notes
+    }
+
     /// Returns an iterator over produced nullifiers for all consumed notes.
     pub fn produced_nullifiers(&self) -> impl Iterator<Item = Nullifier> + '_ {
         self.produced_nullifiers.iter().cloned()
     }
 
-    /// Returns the root hash of the created notes SMT.
-    pub fn created_notes_root(&self) -> Digest {
-        self.created_notes_smt.root()
+    /// Returns the root hash of the output notes SMT.
+    pub fn output_notes_root(&self) -> Digest {
+        self.output_notes_smt.root()
     }
 
-    /// Returns created notes list.
-    pub fn created_notes(&self) -> &Vec<OutputNote> {
-        &self.created_notes
+    /// Returns output notes list.
+    pub fn output_notes(&self) -> &Vec<OutputNote> {
+        &self.output_notes
     }
 
     // HELPER FUNCTIONS