+++ crates/block-producer/src/state_view/mod.rs
@@ -3,7 +3,11 @@ use std::{collections::BTreeSet, sync::Arc};
 use async_trait::async_trait;
 use miden_node_utils::formatting::format_array;
 use miden_objects::{
-    accounts::AccountId, block::Block, notes::Nullifier, Digest, MIN_PROOF_SECURITY_LEVEL,
+    accounts::AccountId,
+    block::Block,
+    notes::{NoteId, Nullifier},
+    transaction::OutputNote,
+    Digest, MIN_PROOF_SECURITY_LEVEL,
 };
 use miden_tx::TransactionVerifier;
 use tokio::sync::RwLock;
@@ -31,6 +35,9 @@ pub struct DefaultStateView<S> {
 
     /// The nullifiers of notes consumed by transactions currently in the block production pipeline.
     nullifiers_in_flight: Arc<RwLock<BTreeSet<Nullifier>>>,
+
+    /// The output notes of transactions currently in the block production pipeline.
+    notes_in_flight: Arc<RwLock<BTreeSet<NoteId>>>,
 }
 
 impl<S> DefaultStateView<S>
@@ -41,8 +48,9 @@ where
         Self {
             store,
             verify_tx_proofs,
-            accounts_in_flight: Arc::new(RwLock::new(BTreeSet::new())),
-            nullifiers_in_flight: Arc::new(RwLock::new(BTreeSet::new())),
+            accounts_in_flight: Default::default(),
+            nullifiers_in_flight: Default::default(),
+            notes_in_flight: Default::default(),
         }
     }
 }
@@ -69,15 +77,23 @@ where
         //
         // This is a "soft" check, because we'll need to redo it at the end. We do this soft check
         // to quickly reject clearly infracting transactions before hitting the store (slow).
+        //
+        // At this stage we don't provide missing notes, they will be available on the second check
+        // after getting the transaction inputs.
         ensure_in_flight_constraints(
             candidate_tx,
             &*self.accounts_in_flight.read().await,
             &*self.nullifiers_in_flight.read().await,
+            &*self.notes_in_flight.read().await,
+            &[],
         )?;
 
-        // Fetch the transaction inputs from the store, and check tx input constraints
+        // Fetch the transaction inputs from the store, and check tx input constraints; this will
+        // identify a set of unauthenticated input notes which are not in the store yet; we'll use
+        // this set to verify that these notes are currently in flight (i.e., they are output notes
+        // of one of the inflight transactions)
         let tx_inputs = self.store.get_tx_inputs(candidate_tx).await?;
-        ensure_tx_inputs_constraints(candidate_tx, tx_inputs)?;
+        let missing_notes = ensure_tx_inputs_constraints(candidate_tx, tx_inputs)?;
 
         // Re-check in-flight transaction constraints, and if verification passes, register
         // transaction
@@ -87,19 +103,20 @@ where
         {
             let mut locked_accounts_in_flight = self.accounts_in_flight.write().await;
             let mut locked_nullifiers_in_flight = self.nullifiers_in_flight.write().await;
+            let mut locked_notes_in_flight = self.notes_in_flight.write().await;
 
             ensure_in_flight_constraints(
                 candidate_tx,
                 &locked_accounts_in_flight,
                 &locked_nullifiers_in_flight,
+                &locked_notes_in_flight,
+                &missing_notes,
             )?;
 
             // Success! Register transaction as successfully verified
             locked_accounts_in_flight.insert(candidate_tx.account_id());
-
-            let mut nullifiers_in_tx: BTreeSet<_> =
-                candidate_tx.input_notes().iter().map(|note| note.nullifier()).collect();
-            locked_nullifiers_in_flight.append(&mut nullifiers_in_tx);
+            locked_nullifiers_in_flight.extend(&mut candidate_tx.get_nullifiers());
+            locked_notes_in_flight.extend(candidate_tx.output_notes().iter().map(OutputNote::id));
         }
 
         Ok(())
@@ -120,6 +137,7 @@ where
 
         let mut locked_accounts_in_flight = self.accounts_in_flight.write().await;
         let mut locked_nullifiers_in_flight = self.nullifiers_in_flight.write().await;
+        let mut locked_notes_in_flight = self.notes_in_flight.write().await;
 
         // Remove account ids of transactions in block
         for update in block.updated_accounts() {
@@ -133,6 +151,14 @@ where
             debug_assert!(was_in_flight);
         }
 
+        // Remove new notes of transactions in block
+        for batch in block.created_notes() {
+            for note in batch.iter() {
+                let was_in_flight = locked_notes_in_flight.remove(&note.id());
+                debug_assert!(was_in_flight);
+            }
+        }
+
         Ok(())
     }
 }
@@ -143,13 +169,16 @@ where
 /// Ensures the constraints related to in-flight transactions:
 /// - the candidate transaction doesn't modify the same account as an existing in-flight
 ///   transaction (issue: #186)
-/// - no consumed note's nullifier in candidate tx's consumed notes is already contained in
+/// - no note's nullifier in candidate tx's consumed notes is already contained in
 ///   `already_consumed_nullifiers`
+/// - all notes in `tx_notes_not_in_store` are currently in flight
 #[instrument(target = "miden-block-producer", skip_all, err)]
 fn ensure_in_flight_constraints(
     candidate_tx: &ProvenTransaction,
     accounts_in_flight: &BTreeSet<AccountId>,
     already_consumed_nullifiers: &BTreeSet<Nullifier>,
+    notes_in_flight: &BTreeSet<NoteId>,
+    tx_notes_not_in_store: &[NoteId],
 ) -> Result<(), VerifyTxError> {
     debug!(target: COMPONENT, accounts_in_flight = %format_array(accounts_in_flight), already_consumed_nullifiers = %format_array(already_consumed_nullifiers));
 
@@ -161,10 +190,8 @@ fn ensure_in_flight_constraints(
     // Check no consumed notes were already consumed
     let infracting_nullifiers: Vec<Nullifier> = {
         candidate_tx
-            .input_notes()
-            .iter()
-            .filter(|&commitment| already_consumed_nullifiers.contains(&commitment.nullifier()))
-            .map(|commitment| commitment.nullifier())
+            .get_nullifiers()
+            .filter(|nullifier| already_consumed_nullifiers.contains(nullifier))
             .collect()
     };
 
@@ -172,14 +199,30 @@ fn ensure_in_flight_constraints(
         return Err(VerifyTxError::InputNotesAlreadyConsumed(infracting_nullifiers));
     }
 
+    // Check all notes not found in store are in in-flight notes, return list of missing notes
+    let missing_notes: Vec<NoteId> = tx_notes_not_in_store
+        .iter()
+        .filter(|note_id| !notes_in_flight.contains(note_id))
+        .copied()
+        .collect();
+    if !missing_notes.is_empty() {
+        return Err(VerifyTxError::UnauthenticatedNotesNotFound(missing_notes));
+    }
+
     Ok(())
 }
 
+/// Ensures the constraints related to transaction inputs:
+/// - the candidate transaction's initial account state hash must be the same as the one
+///   in the Store or empty for new accounts
+/// - input notes must not be already consumed
+///
+/// Returns a list of unauthenticated input notes that were not found in the store.
 #[instrument(target = "miden-block-producer", skip_all, err)]
 fn ensure_tx_inputs_constraints(
     candidate_tx: &ProvenTransaction,
     tx_inputs: TransactionInputs,
-) -> Result<(), VerifyTxError> {
+) -> Result<Vec<NoteId>, VerifyTxError> {
     debug!(target: COMPONENT, %tx_inputs);
 
     match tx_inputs.account_hash {
@@ -209,13 +252,12 @@ fn ensure_tx_inputs_constraints(
     let infracting_nullifiers: Vec<Nullifier> = tx_inputs
         .nullifiers
         .into_iter()
-        .filter(|&(_, block_num)| block_num != 0)
-        .map(|(nullifier_in_tx, _)| nullifier_in_tx)
+        .filter_map(|(nullifier_in_tx, block_num)| (block_num != 0).then_some(nullifier_in_tx))
         .collect();
 
     if !infracting_nullifiers.is_empty() {
         return Err(VerifyTxError::InputNotesAlreadyConsumed(infracting_nullifiers));
     }
 
-    Ok(())
+    Ok(tx_inputs.missing_unauthenticated_notes)
 }