+++ crates/block-producer/src/state_view/tests/apply_block.rs
@@ -6,10 +6,7 @@
 
 use std::iter;
 
-use miden_objects::{
-    accounts::delta::AccountUpdateDetails, block::BlockAccountUpdate,
-    transaction::InputNoteCommitment,
-};
+use miden_objects::{accounts::delta::AccountUpdateDetails, block::BlockAccountUpdate};
 
 use super::*;
 use crate::test_utils::{block::MockBlockBuilder, MockStoreSuccessBuilder};
@@ -161,14 +158,12 @@ async fn test_apply_block_ab3() {
         accounts[0].states[1],
         accounts[0].states[2],
     )
-    .nullifiers(txs[0].input_notes().iter().map(InputNoteCommitment::nullifier).collect())
+    .nullifiers(txs[0].get_nullifiers().collect())
     .build();
 
     let verify_tx_res = state_view.verify_tx(&tx_new).await;
     assert_eq!(
         verify_tx_res,
-        Err(VerifyTxError::InputNotesAlreadyConsumed(
-            txs[0].input_notes().iter().map(InputNoteCommitment::nullifier).collect()
-        ))
+        Err(VerifyTxError::InputNotesAlreadyConsumed(txs[0].get_nullifiers().collect()))
     );
 }